+++ NEWS.md
@@ -1,5 +1,5 @@
 # osprey 0.1.8.9000
-Issues with no news:
+* Added new plot function Adverse Event Category Plot `g_ae_sub`.
 
 # osprey 0.1.8
 ### Miscellaneous